% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinusAddPlot.R
\name{MinusAddPlot}
\alias{MinusAddPlot}
\title{Visualize the amplicon data}
\usage{
MinusAddPlot(data = data, pdf = "Minus_Add_MA_plot.pdf")
}
\arguments{
\item{otuTab}{otu table of your sample}

\item{metaData}{design file}

\item{classToPlot}{which column you want to plot}

\item{topNum}{top n taxa to plot}

\item{col}{colour palette: including all the types of the "display.brewer.all()" in the RColorBrewer package}
}
\value{

}
\description{
Visualize the amplicon data
}
\examples{
otu_table_L2.txt <- system.file("extdata", "otu_table_L2.txt", package = "microVisu")
design.txt <- system.file("extdata", "design.txt", package = "microVisu")
taxBarPlot(otuTab = otu_table_L2.txt, metaData = design.txt,
 classToPlot = "status", topNum = 10, col = "Set3")
}
