% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinusAddPlot.R
\name{MinusAddPlot}
\alias{MinusAddPlot}
\title{Visualize Minus Add Plot}
\usage{
MinusAddPlot(data = GeneExp, pdf = "Minus_Add_MA_plot.pdf")
}
\arguments{
\item{data}{data.frame of sample gene exp.}

\item{pdf}{out-file.}
}
\value{

}
\description{
Visualize Minus Add Plot
}
\examples{
library(data.table)
data(GeneExp)
MinusAddPlot(GeneExp,"Minus_Add_MA_plot.pdf")
}
