% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinusAddPlot.R
\name{MinusAddPlot}
\alias{MinusAddPlot}
\title{Visualize Minus Add Plot}
\usage{
MinusAddPlot(data = data, pdf = "Minus_Add_MA_plot.pdf")
}
\arguments{
\item{data}{data.frame of sample gene exp.}

\item{pdf}{out-file.}
}
\value{

}
\description{
Visualize Minus Add Plot
}
\examples{
library(data.table)
data<-fread(system.file("extdata", "gene_samples_exp.xls", package = "MinusAddPlot"),header=TRUE,stringsAsFactors=FALSE,data.table=FALSE)
pdf("Minus_Add_MA_plot.pdf")
MinusAddPlot(data)
dev.off()
}
